+++ src/network.rs
@@ -1,4 +1,5 @@
 use crate::util::expect_cast;
+use crate::node::*;
 use crate::*;
 
 pub type LayerT = Vec<AnyNode>;
